pyproject.toml
[tool.poetry]
name = "papdl"
version = "0.1.0"
description = ""
authors = ["Tamim Azmain <maat1@st-andrews.ac.uk>"]
readme = "README.md"
packages = [{include = "papdl", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.24.1"
pandas = "^1.5.2"
pyyaml = "^6.0"
tensorflow = "^2.11.0"
keras = "^2.11.0"
asyncio = "^3.4.3"
websockets = "^10.4"
termcolor = "^2.1.1"
crossplane = "^0.5.8"
click = "^8.1.3"
docker = "^6.0.1"
random-word = "^1.0.11"
tqdm = "^4.64.1"
colorama = "^0.4.6"
pyopenssl = "^23.0.0"
iperf3 = "^0.1.11"
pythonping = "^1.1.4"
jsonpickle = "^3.0.1"
autopep8 = "^2.0.1"
sanic = "^22.12.0"
asynchttp = "^0.0.4"
aiohttp = "^3.8.4"
uproot = "^5.0.4"
matplotlib = "^3.7.1"
ipykernel = "^6.21.3"
notebook = "^6.5.3"
psutil = "^5.9.4"
coloredlogs = "^15.0.1"
dill = "^0.3.6"
asyncstdlib = "^3.10.5"

[tool.poetry.scripts]
papdl = "papdl.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
